requirements.txt
absl-py==2.1.0
addict==2.4.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bleach==6.1.0
blobconverter==1.4.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
depthai==2.21.2.0
depthai-sdk==1.9.4
distinctipy==1.3.4
executing==2.0.1
fastjsonschema==2.19.1
ffmpy3==0.2.4
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.52.4
fqdn==1.5.1
fsspec==2024.6.0
future==1.0.0
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.3.post0
lmdb==1.4.1
lz4==4.3.3
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.17.0
matplotlib==3.9.0
matplotlib-inline==0.1.7
mcap==1.1.1
mcap-ros1-support==0.0.8
mistune==3.0.2
mkl==2021.4.0
mmcv==0.6.2
mmcv-full==1.7.0
mmdet==2.27.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.82
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pycocotools==2.0.7
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytube==15.0.0
PyTurboJPEG==1.6.4
pytz==2024.1
pyusb==1.2.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
Qt.py==1.3.10
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rosbags==0.9.11
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-image==0.23.2
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==1.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12.1
tb-nightly==2.18.0a20240616
tbb==2021.12.0
tensorboard-data-server==0.7.2
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.5.22
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.1+cu121
torch-pruning==1.4.1
torchmetrics==1.4.0.post0
torchvision==0.18.1+cu121
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-pyside2==5.15.2.1.6
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
xmltodict==0.13.0
yapf==0.40.2
zipp==3.19.2
